+++ src/main/java/com/microej/kernel/green/security/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Java
+ *
+ * Copyright 2023 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ */
+/**
+ * This package contains security-related classes.
+ */
+package com.microej.kernel.green.security;