+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.0] - 2023-10-27
+
+### Added
+
+- Add logging-only SecurityManager.
+- Add kernel.intern file for NET module usage.
+
+### Changed
+
+- Update the service, property and kf-util dependencies and usages.
+- Update documentation.
+- Change the default way of fetching a VEE Port (switch from remote module dependency to local source)
+
 ## [1.1.0] - 2023-09-13
 
 ### Added